Refactoring of older code done
