+++ src/components/Chip/Chip.tsx
@@ -3,6 +3,10 @@ import styled from 'styled-components';
 import React from 'react';
 import * as cssVariables from '../CommonTextStyles/variables';
 
+/**
+ * Creates a selectable chip
+ */
+
 const Chip: React.FC<ChipProps> = ({ text }) => {
   const [isChipSelected, setChipSelected] = React.useState(false);
