+++ src/features/MentorPage/MentorChips/MentorChips.test.tsx
@@ -6,11 +6,6 @@ import { BrowserRouter } from 'react-router-dom';
 import { ChipProps } from '../../../components/Chip/types';
 import MentorChips from './MentorChips';
 
-/**
- * When functionality is added, a test should be written that
- * checks that right chips are shown
- */
-
 describe('<MentorChips />', () => {
   const items: Array<ChipProps> = [{ text: 'test chip' }];
