--- src/features/MentorSearch/index.ts
@@ -1,3 +1,0 @@
-import MentorSearch from './MentorSearch';
-
-export default MentorSearch;
