--- src/features/MentorSkills/index.ts
@@ -1,3 +1,0 @@
-import MentorSkills from './MentorSkills';
-
-export default MentorSkills;
